# src/app/wwwroot/swagger/helium/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Helium (C#)",
    "version": "1.0.1.0"
  },
  "paths": {
...
    "/api/Movies": {
      "get": {
        "tags": [ "Movies" ],
        "summary": "Search for Movies",
        "description": "Returns a JSON array of Movie documents",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "The term used to search by Movie title <br> ```valid: 3 - 50 characters```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "matrix"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "description": "Returns a JSON array of Movie documents filtered by genre<br> ```valid: 3 - 50 characters```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "action"
            }
          },
          {
            "name": "year",
            "in": "query",
            "description": "Get movies by year <br>\n  ```valid: 1873 - currentYear + 5```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "example": 2003
            }
          },
          {
            "name": "rating",
            "in": "query",
            "description": "Get movies with a rating >= rating <br>\n  ```valid: 0.0 - 10.0```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "number",
              "example": 6.5
            }
          },
          {
            "name": "actorId",
            "in": "query",
            "description": "Get movies by Actor Id <br>\n  ```valid: nm0000206```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "nm0000206"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "1 based page index <br>\n  ```valid: 1 - 10,000```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "example": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "page size <br>\n  ```valid 1 - 1000```",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "example": 10,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON array of Movie documents or empty array if not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                },
                "example": [
                  {
                    "id": "tt0234215",
                    "partitionKey": "5",
                    "movieId": "tt0234215",
                    "type": "Movie",
                    "title": "The Matrix Reloaded",
                    "year": 2003,
                    "runtime": 138,
                    "rating": 7.2,
                    "votes": 486680,
                    "totalScore": 3504096,
                    "textSearch": "the matrix reloaded",
                    "genres": [
                      "Action",
                      "Sci-Fi"
                    ],
                    "roles": [
                      {
                        "order": 1,
                        "actorId": "nm0000206",
                        "name": "Keanu Reeves",
                        "birthYear": 1964,
                        "category": "Actor",
                        "characters": [
                          "Neo"
                        ]
                      },
                      {
                        "order": 2,
                        "actorId": "nm0000401",
                        "name": "Laurence Fishburne",
                        "birthYear": 1961,
                        "category": "Actor",
                        "characters": [
                          "Morpheus"
                        ]
                      },
                      {
                        "order": 3,
                        "actorId": "nm0005251",
                        "name": "Carrie-Anne Moss",
                        "birthYear": 1967,
                        "category": "Actress",
                        "characters": [
                          "Trinity"
                        ]
                      },
                      {
                        "order": 4,
                        "actorId": "nm0915989",
                        "name": "Hugo Weaving",
                        "birthYear": 1960,
                        "category": "Actor",
                        "characters": [
                          "Agent Smith"
                        ]
                      },
                      {
                        "order": 5,
                        "actorId": "nm0905154",
                        "name": "Lana Wachowski",
                        "birthYear": 1965,
                        "category": "Director"
                      },
                      {
                        "order": 6,
                        "actorId": "nm0905152",
                        "name": "Lilly Wachowski",
                        "birthYear": 1967,
                        "category": "Director"
                      },
                      {
                        "order": 7,
                        "actorId": "nm0005428",
                        "name": "Joel Silver",
                        "birthYear": 1952,
                        "category": "Producer"
                      }
                    ]
                  },
                  {
                    "id": "tt0242653",
                    "partitionKey": "3",
                    "movieId": "tt0242653",
                    "type": "Movie",
                    "title": "The Matrix Revolutions",
                    "year": 2003,
                    "runtime": 129,
                    "rating": 6.7,
                    "votes": 422540,
                    "totalScore": 2831018,
                    "textSearch": "the matrix revolutions",
                    "genres": [
                      "Action",
                      "Sci-Fi"
                    ],
                    "roles": [
                      {
                        "order": 1,
                        "actorId": "nm0000206",
                        "name": "Keanu Reeves",
                        "birthYear": 1964,
                        "category": "Actor",
                        "characters": [
                          "Neo"
                        ]
                      },
                      {
                        "order": 2,
                        "actorId": "nm0000401",
                        "name": "Laurence Fishburne",
                        "birthYear": 1961,
                        "category": "Actor",
                        "characters": [
                          "Morpheus"
                        ]
                      },
                      {
                        "order": 3,
                        "actorId": "nm0005251",
                        "name": "Carrie-Anne Moss",
                        "birthYear": 1967,
                        "category": "Actress",
                        "characters": [
                          "Trinity"
                        ]
                      },
                      {
                        "order": 4,
                        "actorId": "nm0915989",
                        "name": "Hugo Weaving",
                        "birthYear": 1960,
                        "category": "Actor",
                        "characters": [
                          "Agent Smith"
                        ]
                      },
                      {
                        "order": 5,
                        "actorId": "nm0905154",
                        "name": "Lana Wachowski",
                        "birthYear": 1965,
                        "category": "Director"
                      },
                      {
                        "order": 6,
                        "actorId": "nm0905152",
                        "name": "Lilly Wachowski",
                        "birthYear": 1967,
                        "category": "Director"
                      },
                      {
                        "order": 7,
                        "actorId": "nm0005428",
                        "name": "Joel Silver",
                        "birthYear": 1952,
                        "category": "Producer"
                      }
                    ]
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Invalid search parameter"
                }
              }
            }
          }
        }
      }
    },
    "/api/Movies/{movieId}": {
      "get": {
        "tags": [ "Movies" ],
        "summary": "Get a Movie by ID",
        "description": "Returns a single JSON Movie by movieId",
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "example - tt0133093 - The Matrix",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            },
            "example": "tt0133093"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                },
                "example": {
                  "id": "tt0133093",
                  "partitionKey": "3",
                  "movieId": "tt0133093",
                  "type": "Movie",
                  "title": "The Matrix",
                  "year": 1999,
                  "runtime": 136,
                  "rating": 8.7,
                  "votes": 1546282,
                  "totalScore": 13452653,
                  "genres": [ "Action", "Sci-Fi" ],
                  "roles": [
                    {
                      "order": 1,
                      "actorId": "nm0000206",
                      "name": "Keanu Reeves",
                      "birthYear": 1964,
                      "category": "Actor",
                      "characters": [ "Neo" ]
                    },
                    {
                      "order": 2,
                      "actorId": "nm0000401",
                      "name": "Laurence Fishburne",
                      "birthYear": 1961,
                      "category": "Actor",
                      "characters": [ "Morpheus" ]
                    },
                    {
                      "order": 3,
                      "actorId": "nm0005251",
                      "name": "Carrie-Anne Moss",
                      "birthYear": 1967,
                      "category": "Actress",
                      "characters": [ "Trinity" ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid movieId",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Invalid movieId"
                }
              }
            }
          },
          "404": {
            "description": "movieId not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "movieId not found"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "partitionKey": {
            "type": "string"
          },
          "movieId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "runtime": {
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "votes": {
            "type": "integer"
          },
          "totalScore": {
            "type": "integer"
          },
          "textSearch": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "order": {
                  "type": "integer"
                },
                "actorId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "birthYear": {
                  "type": "integer"
                },
                "deathYear": {
                  "type": "integer",
                  "nullable": true
                },
                "category": {
                  "type": "string"
                },
                "characters": {
                  "type": "array",
                  "nullable": true,
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}